+++ servers/view/process.php
@@ -6,35 +6,11 @@ $response = array();
 // Server side check to make sure nothing is blank / empty
 if(!empty($_POST))
 {
-  require_once("../../classes/class.Player.inc.php");
+  require_once("../../classes/class.Server.inc.php");
 
-	$player = new \WurmUnlimitedAdmin\PLAYER();
+	$server = new \WurmUnlimitedAdmin\SERVER();
 	switch($_POST["doing"])
 	{
-		case "banFunction":
-			$response = $player->BanUnban($_POST);
-			break;
-		case "muteFunction":
-			$response = $player->MuteUnmute($_POST);
-			break;
-		case "changePower":
-			$response = $player->ChangePower($_POST);
-			break;
-		case "addMoney":
-			$response = $player->AddMoney($_POST);
-			break;
-		case "changeEmail":
-			$response = $player->ChangeEmail($_POST);
-			break;
-		case "changeKingdom":
-			$response = $player->ChangeKingdom($_POST);
-			break;
-		case "getInventory":
-			$response = $player->GetInventory($_POST["playerID"]);
-			break;
-		case "getSkills":
-			$response = $player->GetSkills($_POST["playerID"]);
-			break;
 		default:
 			$response = array("success" => false);
 			break;